Made specific file for plotting
